Tests\Modules\History: Change - improve performance.
